# docs/admin/swagger.yaml
definitions:
  enum.EProductType:
    enum:
    - simple
    - variant
    - bundle
    - admin
    - customer
    type: string
    x-enum-varnames:
    - ProductTypeSimple
    - ProductTypeVariant
    - ProductTypeGroup
    - Admin
    - Customer
...
  request.ProductCategory:
    properties:
      name:
        type: string
      parentId:
        type: integer
    type: object
  request.ProductGroupItemPayload:
    properties:
      id:
        type: string
      productId:
        type: string
      productSkuId:
        type: string
      qty:
        type: integer
    type: object
  request.ProductPayload:
    properties:
      description:
        type: string
      imageIds:
        items:
          type: string
        type: array
      isActive:
        type: boolean
      isShowOnlyInMarketPlace:
        type: boolean
      name:
        type: string
      productCategoryId:
        type: integer
      productGroupItems:
        items:
          $ref: '#/definitions/request.ProductGroupItemPayload'
        type: array
      productSkuVariants:
        items:
          $ref: '#/definitions/request.ProductSkuVariantPayload'
        type: array
      productSkus:
        items:
          $ref: '#/definitions/request.ProductSkuPayload'
        type: array
      productType:
        $ref: '#/definitions/enum.EProductType'
      productVariantOptionValues:
        items:
          $ref: '#/definitions/request.ProductVariantOptionValuePayload'
        type: array
      productVariantOptions:
        items:
          $ref: '#/definitions/request.ProductVariantOptionPayload'
        type: array
      sku:
        type: string
    type: object
  request.ProductSkuPayload:
    properties:
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      price:
        type: number
      productSpecs:
        items:
          $ref: '#/definitions/request.ProductSpec'
        type: array
      sku:
        type: string
      stock:
        type: integer
      stockAlert:
        type: integer
    type: object
  request.ProductSkuVariantPayload:
    properties:
      id:
        type: string
      productSkuId:
        type: string
      productVariantOptionId:
        type: string
      productVariantOptionValueId:
        type: string
    type: object
  request.ProductSpec:
    properties:
      id:
        type: string
      specKey:
        type: string
      specValue:
        type: string
    type: object
  request.ProductVariantOptionPayload:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  request.ProductVariantOptionValuePayload:
    properties:
      id:
        type: string
      name:
        type: string
      productVariantOptionId:
        type: string
    type: object
...
  request.VariantOptionPayload:
    properties:
      name:
        type: string
    type: object
  response.AuditTrail:
    properties:
      id:
        type: integer
      userName:
        type: string
    type: object
...
  response.EnumResponse:
    properties:
      code:
        type: string
      name:
        type: string
    type: object
  response.ProductCategory:
    properties:
      createdAt:
        type: string
      createdById:
        $ref: '#/definitions/response.AuditTrail'
      id:
        type: integer
      modifiedAt:
        type: string
      modifiedById:
        $ref: '#/definitions/response.AuditTrail'
      name:
        type: string
      parentId:
        type: integer
    type: object
  response.ProductGroupItemResponse:
    properties:
      id:
        type: string
      product:
        $ref: '#/definitions/response.ProductResponse'
      productId:
        type: string
      productSkuId:
        type: string
      qty:
        type: integer
    type: object
  response.ProductResponse:
    properties:
      createdAt:
        type: string
      createdById:
        type: integer
      description:
        type: string
      id:
        type: string
      imageIds:
        items:
          type: string
        type: array
      isActive:
        type: boolean
      isShowOnlyInMarketPlace:
        type: boolean
      modifiedAt:
        type: string
      modifiedById:
        type: integer
      name:
        type: string
      productCategory:
        $ref: '#/definitions/response.ProductCategory'
      productGroupItems:
        items:
          $ref: '#/definitions/response.ProductGroupItemResponse'
        type: array
      productSkuVariants:
        items:
          $ref: '#/definitions/response.ProductSkuVariant'
        type: array
      productSkus:
        items:
          $ref: '#/definitions/response.ProductSku'
        type: array
      productType:
        $ref: '#/definitions/response.EnumResponse'
      productVariantOptionValues:
        items:
          $ref: '#/definitions/response.ProductVariantOptionValue'
        type: array
      productVariantOptions:
        items:
          $ref: '#/definitions/response.ProductVariantOption'
        type: array
      sku:
        type: string
    type: object
  response.ProductSku:
    properties:
      ProductSpecs:
        items:
          $ref: '#/definitions/response.ProductSpec'
        type: array
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      price:
        type: number
      productId:
        type: string
      sequence:
        type: integer
      sku:
        type: string
      stock:
        type: integer
      stockAlert:
        type: integer
    type: object
  response.ProductSkuVariant:
    properties:
      id:
        type: string
      productSkuId:
        type: string
      productVariantOptionId:
        type: string
      productVariantOptionValueId:
        type: string
    type: object
  response.ProductSpec:
    properties:
      id:
        type: string
      productSkuId:
        type: string
      specKey:
        type: string
      specValue:
        type: string
    type: object
  response.ProductVariantOption:
    properties:
      allowDelete:
        type: boolean
      id:
        type: string
      name:
        type: string
      productId:
        type: string
      sequence:
        type: integer
    type: object
  response.ProductVariantOptionValue:
    properties:
      allowDelete:
        type: boolean
      id:
        type: string
      name:
        type: string
      productVariantOptionId:
        type: string
      sequence:
        type: integer
    type: object
...
  response.VariantOption:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  response.WebResponse:
    properties:
      code:
        type: integer
      content: {}
      message:
        items:
          type: string
        type: array
      status:
        type: string
    type: object
info:
  contact: {}
  description: Swagger documentation for admin
  title: Admin API
  version: "1.0"
paths:
...
  /api/admin/product-categories:
    get:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: query
        name: query
        type: string
      - description: ' '
        in: query
        name: parentId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductCategory'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Search category
      tags:
      - Product Category
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ProductCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Create category
      tags:
      - Product Category
  /api/admin/product-categories/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete category
      tags:
      - Product Category
    get:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductCategory'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Get category by id
      tags:
      - Product Category
    put:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      - description: ' '
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ProductCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Update category
      tags:
      - Product Category
  /api/admin/products:
    get:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: query
        name: query
        type: string
      - description: ' '
        in: query
        name: productTypes
        type: string
      - description: ' '
        in: query
        name: productCategoryId
        type: integer
      - description: ' '
        in: query
        name: isActive
        type: boolean
      - description: ' '
        in: query
        name: isShowOnlyInMarketplace
        type: boolean
      - description: ' '
        in: query
        name: page
        type: integer
      - description: ' '
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ProductResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Search products
      tags:
      - Product
  /api/admin/products/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete product
      tags:
      - Product
    get:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Get product by id
      tags:
      - Product
    post:
      consumes:
      - application/json
      parameters:
      - description: Id (UUID format)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: ' '
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ProductPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Product
      tags:
      - Product
    put:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: string
      - description: ' '
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ProductPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Product
      tags:
      - Product
  /api/admin/products/browse-product-sku:
    get:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: query
        name: query
        type: string
      - description: ' '
        in: query
        name: productTypes
        type: string
      - description: ' '
        in: query
        name: productCategoryId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ProductResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Browse products sku
      tags:
      - Product
  /api/admin/variant-options:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.VariantOption'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Variant Options
      tags:
      - Product Variant Option
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.VariantOptionPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Variant Option
      tags:
      - Product Variant Option
...
swagger: "2.0"